nklfehkeh

+++ app/Http/Controllers/StaffController.php
@@ -12,6 +12,8 @@ class StaffController extends Controller
     // Display a listing of the resource.
     public function index()
     {
+        $this->authorize('viewAny', Staff::class);
+
         $staffs = Staff::all();
         return view('admin.staff.index', compact('staffs'));
     }
@@ -19,6 +21,8 @@ class StaffController extends Controller
     // Show the form for creating a new resource.
     public function create()
     {
+        $this->authorize('create', Staff::class);
+
         $managers = User::where('role', 'manager')->get();
         $users = User::where('role', 'staff')->get();
         return view('admin.staff.create', compact([
@@ -30,6 +34,8 @@ class StaffController extends Controller
     // Store a newly created resource in storage.
     public function store(Request $request)
     {
+        $this->authorize('create', Staff::class);
+
         $request->validate([
             'user_id' => 'required|exists:users,id',
             'manager_id' => 'nullable|exists:users,id',
@@ -45,6 +51,8 @@ class StaffController extends Controller
     // Show the form for editing the specified resource.
     public function edit(Staff $staff)
     {
+        $this->authorize('update', $staff);
+
         $managers = User::where('role', 'manager')->get();
         $users = User::where('role', 'staff');
         return view('admin.staff.edit', compact(
@@ -57,6 +65,8 @@ class StaffController extends Controller
     // Update the specified resource in storage.
     public function update(Request $request, Staff $staff)
     {
+        $this->authorize('update', $staff);
+
         $request->validate([
             'user_id' => 'required|exists:users,id',
             'manager_id' => 'nullable|exists:users,id',
@@ -72,6 +82,7 @@ class StaffController extends Controller
     // Remove the specified resource from storage.
     public function destroy(Staff $staff)
     {
+        $this->authorize('delete', $staff);
         $staff->delete();
 
         return redirect()->route('staffs.index')->with('success', 'Staff deleted successfully.');
